Moved bookkeeper to separate module
